+++ st2actions/st2actions/policies/concurrency.py
@@ -77,8 +77,7 @@ class ConcurrencyApplicator(BaseConcurrencyApplicator):
 
         valid_states = [
             action_constants.LIVEACTION_STATUS_REQUESTED,
-            action_constants.LIVEACTION_STATUS_DELAYED,
-            action_constants.LIVEACTION_STATUS_POLICY_DELAYED,
+            action_constants.LIVEACTION_STATUS_DELAYED
         ]
 
         # Exit if target not in valid state.